update map url param

+++ build.sh
@@ -1,7 +1,5 @@
 #!/usr/bin/env bash
 
-set -e
-
 LINT_EXIT_CODE=0
 
 eslint public/scripts/*.js
@@ -36,11 +34,11 @@ cp  public/*.xml dist/
 cp  public/site.webmanifest dist/
 
 # simple hash on the JS file for cache busting
-[[ $TRAVIS_BUILD_NUMBER ]] && cache_hash=-$TRAVIS_BUILD_NUMBER
+[[ $TRAVIS_BUILD_NUMBER ]] && cache_hash=_$TRAVIS_BUILD_NUMBER
 index_file=index$cache_hash.min.js
-uglifyjs -cm toplevel --verbose --warn ./public/scripts/**/*.js ./public/scripts/*.js -o ./dist/$index_file --source-map url=index.min.js.map
-# if [ $? -ne 0 ]; then
-#     exit 1
-# fi
+uglifyjs -cm toplevel --verbose --warn ./public/scripts/**/*.js ./public/scripts/*.js -o ./dist/$index_file --source-map url="$index_file.map"
+if [ $? -ne 0 ]; then
+    exit 1
+fi
 sed -i.bak 's/{{index_file}}/'"${index_file}"'/' dist/index.html && rm dist/index.html.bak
 # cp public/scripts/polyfill/codepointat.js public/scripts/index.js dist/
